requirements.txt
requests
pandas
python-dotenv
openpyxl
pycountry
openai
find-job-titles
langdetect
PyQt5
fastapi
uvicorn
python-multipart
streamlit
Pillow